WIP: Configurable Worker Signal Handling
The goal is to make Unicorn's behavior of QUIT and TERM compliant with standard signal handling. Similar to the efforts that were taken in Resque: http://hone.heroku.com/resque/2012/08/21/resque-signals.html

I was able to isolate the behavior to a failing test, then make it pass in almost the simplest way possible.

run tests with

```
$ make test/unit/test_signals.rb--test_workers_finish_processing_requests_after_term.n
```

For more info on running tests see the HACKING file. Substitute `make` for `gmake` if using a mac.